requirements.txt
requests>=2.25.1

# Core numeric stack (pin to avoid pandas 3.0 breakage)
numpy>=1.24,<2.5
pandas>=2.2.2,<3
scipy>=1.11,<1.18

# Data sources
pandas_datareader==0.10.0
yfinance>=0.2.0,<2.0
yahoo_fin>=0.8.9.1

# Parsing / utilities
beautifulsoup4>=4.9.3
PyYAML>=5.4.1
lxml
requests-html
